SDK-4363 Enable and fix sign-conversion warning for SDK library CMake target in Clang

Closes SDK-4363

See merge request sdk/sdk!6002

+++ src/file.cpp
@@ -301,7 +301,7 @@ File *File::unserialize(string *d)
         }
     }
 
-    d->erase(0, ptr - d->data());
+    d->erase(0, static_cast<size_t>(ptr - d->data()));
     return file;
 }
